refactor : adapter에 속한 ItemModel, ViewHolder를 해당 하위 폴더로 위치 시킴
