+++ app/src/main/java/woowacourse/movie/ui/adapter/ReservationListAdapter.kt
@@ -4,10 +4,10 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
-import woowacourse.movie.ui.itemModel.ItemModel
-import woowacourse.movie.ui.itemModel.TicketsItemModel
-import woowacourse.movie.ui.viewHolder.ItemViewHolder
-import woowacourse.movie.ui.viewHolder.TicketsViewHolder
+import woowacourse.movie.ui.adapter.itemModel.ItemModel
+import woowacourse.movie.ui.adapter.itemModel.TicketsItemModel
+import woowacourse.movie.ui.adapter.viewHolder.ItemViewHolder
+import woowacourse.movie.ui.adapter.viewHolder.TicketsViewHolder
 
 class ReservationListAdapter(
     reservations: List<ItemModel>,
