+++ app/src/main/java/woowacourse/movie/ui/adapter/itemModel/AdvItemModel.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.ui.itemModel
+package woowacourse.movie.ui.adapter.itemModel
 
 import woowacourse.movie.model.AdvState
