+++ app/src/main/java/woowacourse/movie/ui/adapter/itemModel/ItemModel.kt
@@ -1,0 +1,3 @@
+package woowacourse.movie.ui.adapter.itemModel
+
+interface ItemModel
