--- app/src/main/java/woowacourse/movie/ui/itemModel/ItemModel.kt
@@ -1,3 +1,0 @@
-package woowacourse.movie.ui.itemModel
-
-interface ItemModel
